clean modal files
